Create reusable ledger book transactions component
