+++ adminapp/src/components/PaymentAccountRelatedLists.jsx
@@ -1,7 +1,8 @@
 import { dayjs } from "../modules/dayConfig";
-import { formatMoney, scaleMoney } from "../shared/money";
+import { formatMoney } from "../shared/money";
 import Money from "../shared/react/Money";
 import AdminLink from "./AdminLink";
+import LedgerBookTransactionsRelatedList from "./LedgerBookTransactionRelatedList";
 import Link from "./Link";
 import RelatedList from "./RelatedList";
 import first from "lodash/first";
@@ -40,39 +41,11 @@ export default function PaymentAccountRelatedLists({ paymentAccount }) {
         ]}
       />
       {paymentAccount.ledgers.map((ledger) => (
-        <RelatedList
+        <LedgerBookTransactionsRelatedList
+          ledger={ledger}
           title={`Ledger ${ledger.label} (${ledger.id}) - ${formatMoney(ledger.balance)}`}
           key={ledger.id}
-          headers={[
-            "Id",
-            "Created",
-            "Applied",
-            "Amount",
-            "Actor",
-            "Category",
-            "Originating",
-            "Receiving",
-          ]}
           rows={ledger.combinedBookTransactions}
-          keyRowAttr="id"
-          toCells={(row) => [
-            <AdminLink key="id" model={row} />,
-            dayjs(row.createdAt).format("lll"),
-            dayjs(row.applyAt).format("lll"),
-            <Money key="amt" accounting>
-              {row.originatingLedger.id === ledger.id
-                ? scaleMoney(row.amount, -1)
-                : row.amount}
-            </Money>,
-            <AdminLink model={row.actor}>{row.actor?.name}</AdminLink>,
-            row.associatedVendorServiceCategory?.name,
-            <AdminLink key="originating" model={row.originatingLedger}>
-              {row.originatingLedger.adminLabel}
-            </AdminLink>,
-            <AdminLink key="receiving" model={row.receivingLedger}>
-              {row.receivingLedger.adminLabel}
-            </AdminLink>,
-          ]}
         />
       ))}
       <RelatedList
