+++ adminapp/src/pages/PlatformLedgerDetailPage.jsx
@@ -1,10 +1,9 @@
 import api from "../api";
-import AdminLink from "../components/AdminLink";
 import DetailGrid from "../components/DetailGrid";
+import LedgerBookTransactionsRelatedList from "../components/LedgerBookTransactionRelatedList";
 import RelatedList from "../components/RelatedList";
 import useErrorSnackbar from "../hooks/useErrorSnackbar";
 import { dayjs } from "../modules/dayConfig";
-import { scaleMoney } from "../shared/money";
 import Money from "../shared/react/Money";
 import useAsyncFetch from "../shared/react/useAsyncFetch";
 import { CircularProgress } from "@mui/material";
@@ -46,36 +45,10 @@ export default function PlatformLedgerDetailPage() {
             keyRowAttr="id"
             toCells={(row) => [row.id, row.name, row.slug]}
           />
-          <RelatedList
+          <LedgerBookTransactionsRelatedList
+            ledger={ledger}
             title="Book Transactions"
-            headers={[
-              "Id",
-              "Created",
-              "Applied",
-              "Amount",
-              "Category",
-              "Originating",
-              "Receiving",
-            ]}
             rows={ledger.combinedBookTransactions}
-            keyRowAttr="id"
-            toCells={(row) => [
-              <AdminLink key="id" model={row} />,
-              dayjs(row.createdAt).format("lll"),
-              dayjs(row.applyAt).format("lll"),
-              <Money key="amt" accounting>
-                {row.originatingLedger.id === ledger.id
-                  ? scaleMoney(row.amount, -1)
-                  : row.amount}
-              </Money>,
-              row.associatedVendorServiceCategory?.name,
-              <AdminLink key="originating" model={row.originatingLedger}>
-                {row.originatingLedger.adminLabel}
-              </AdminLink>,
-              <AdminLink key="receiving" model={row.receivingLedger}>
-                {row.receivingLedger.adminLabel}
-              </AdminLink>,
-            ]}
           />
         </>
       )}
